+++ Misc/Container.h
@@ -4,9 +4,12 @@
  * authors: Tomasz Mrugalski <thomson@klub.com.pl>
  *          Marek Senderski <msend@o2.pl>
  *
- * $Id: Container.h,v 1.4 2004-10-25 20:45:54 thomson Exp $
+ * $Id: Container.h,v 1.5 2005-07-17 21:09:52 thomson Exp $
  *
  * $Log: not supported by cvs2svn $
+ * Revision 1.4  2004/10/25 20:45:54  thomson
+ * Option support, parsers rewritten. ClntIfaceMgr now handles options.
+ *
  * Revision 1.3  2004/03/29 22:06:49  thomson
  * 0.1.1 version
  *
@@ -63,7 +66,7 @@ void TContainer<TYP>::clear() {
 
 template <class TYP>
 int TContainer<TYP>::count() {
-	return lista.size();
+	return (int)lista.size();
 }
 
 template <class TYP>
